+++ cmd/kylix/main.go
@@ -7,7 +7,7 @@ import (
 	"os"
 )
 
-const Version = "0.6.9"
+const Version = "0.7.0"
 
 func main() {
 	if len(os.Args) < 2 {